backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
audioread==3.0.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
distro==1.9.0
fastapi==0.115.0
google-api-core==2.20.0
google-auth==2.35.0
google-cloud-speech==2.27.0
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.5.0
joblib==1.4.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
msgpack==1.1.0
numba==0.60.0
numpy==2.0.2
openai==1.51.0
packaging==24.1
platformdirs==4.3.3
pooch==1.8.2
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
python-dotenv==1.0.1
python-multipart==0.0.12
requests==2.32.3
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
sniffio==1.3.1
soundfile==0.12.1
soxr==0.5.0.post1
starlette==0.38.6
threadpoolctl==3.5.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.0
websockets==13.1
boto3==1.35.0
pyjwt==2.9.0
cryptography==42.0.0
hypothesis==6.92.0
pytest==7.4.3
pytest-asyncio==0.21.1
aiomysql==0.2.0
bcrypt==4.1.2
google-genai==0.3.0
